Refactor gameplay config loading helpers
